django/requirements.txt
python-dotenv
replicate
gunicorn
whitenoise
pymupdf4llm
gspread
google-auth
google-auth-oauthlib
Django
djangorestframework
django-storages
boto3
